pyproject.toml
[project]
name = "slategray-blender-tools"
version = "1.0.0"
description = "A collection of high-performance Blender tools developed by Slategray"
authors = [
    { name = "Randall Rosas (Slategray)" }
]
requires-python = ">=3.10"
dependencies = [
    "numpy",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "basedpyright",
    "fake-bpy-module-latest",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "D",   # pydocstyle
    "PL",  # pylint
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "N801", # Class name convention (Blender uses SBT_OT_...)
    "N806", # Variable in function name (Blender operators use self.layout)
    "PLR0913", # Too many arguments
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.basedpyright]
typeCheckingMode = "basic"
pythonVersion = "3.11"
reportMissingTypeStubs = false
reportUnusedImport = "warning"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
useLibraryCodeForTypes = true
